Move ContinuousTrigger to Triggers with deprecating legacy, deduplicate codes...
